+++ package.json
@@ -9,6 +9,7 @@
     "lint": "turbo run lint",
     "lint:fix": "turbo run lint -- --fix",
     "test": "turbo run test",
+    "test:int": "turbo run test:int",
     "test:coverage": "turbo run test:coverage",
     "test:coverage:html": "turbo run test:coverage:html",
     "coverage:badges": "turbo run coverage:badges && istanbul-badges-readme --coverageDir=./packages/trading/coverage --readmeDir=. --exitCode=1",
