+++ packages/bridging/package.json
@@ -23,11 +23,12 @@
     "lint": "eslint src/**/*.ts",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
     "test": "jest",
+    "test:int": "jest --testMatch='**/*.int.ts' --runInBand",
     "test:coverage": "jest --coverage",
     "coverage:badges": "istanbul-badges-readme --exitCode=1",
     "test:coverage:html": "jest --silent=false --coverage --coverageReporters html",
-    "test:bungeeGnosisBridge": "jest --testPathPattern=BungeeGnosisBridge",
-    "test:bungeeAcrossBridge": "jest --testPathPattern=BungeeAcrossBridge",
+    "test:bungeeGnosisBridge": "jest --testMatch='**/BungeeGnosisBridge.int.ts' --runInBand",
+    "test:bungeeAcrossBridge": "jest --testMatch='**/BungeeAcrossBridge.int.ts' --runInBand",
     "typecheck": "tsc --noEmit",
     "prepublishOnly": "npm run build"
   },
