+++ packages/composable/package.json
@@ -24,6 +24,7 @@
     "lint": "eslint src/**/*.ts",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
     "test": "jest",
+    "test:int": "jest --testMatch='**/*.int.ts' --runInBand",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
     "coverage:badges": "istanbul-badges-readme --exitCode=1",
