+++ packages/cow-shed/package.json
@@ -22,6 +22,7 @@
     "build": "tsup src/index.ts --format esm,cjs --dts",
     "lint": "eslint src/**/*.ts --quiet",
     "test": "jest",
+    "test:int": "jest --testMatch='**/*.int.ts' --runInBand",
     "test:coverage": "jest --coverage",
     "coverage:badges": "istanbul-badges-readme --exitCode=1",
     "test:coverage:html": "jest --silent=false --coverage --coverageReporters html",
