+++ packages/flash-loans/package.json
@@ -23,12 +23,13 @@
     "lint": "eslint src/**/*.ts",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
     "test": "jest",
+    "test:int": "jest --testMatch='**/*.int.ts' --runInBand",
     "test:coverage": "jest --coverage",
     "coverage:badges": "istanbul-badges-readme --exitCode=1",
     "test:coverage:html": "jest --silent=false --coverage --coverageReporters html",
-    "test:integration:aave-collateral-swap": "npx jest --testNamePattern=^AaveFlashLoanIntegration.collateralSwap",
-    "test:integration:aave-debt-swap": "npx jest --testNamePattern=^AaveFlashLoanIntegration.debtSwap",
-    "test:integration:aave-repay-collateral": "npx jest --testNamePattern=^AaveFlashLoanIntegration.repayCollateral",
+    "test:integration:aave-collateral-swap": "jest --testMatch='**/collateralSwap.int.ts' --runInBand",
+    "test:integration:aave-debt-swap": "jest --testMatch='**/debtSwap.int.ts' --runInBand",
+    "test:integration:aave-repay-collateral": "jest --testMatch='**/repayCollateral.int.ts' --runInBand",
     "typecheck": "tsc --noEmit",
     "prepublishOnly": "npm run build"
   },
