+++ turbo.json
@@ -14,6 +14,11 @@
       "dependsOn": ["build"],
       "outputs": []
     },
+    "test:int": {
+      "dependsOn": ["build"],
+      "outputs": [],
+      "cache": false
+    },
     "test:coverage": {
       "outputs": ["coverage/**"],
       "dependsOn": ["build"],
